Improved operation interceptor instantiation (not yet merged upstream) (#210)
