+++ hbase-client/src/main/java/org/apache/hadoop/hbase/client/ConnectionImplementation.java
@@ -257,8 +257,6 @@ public class ConnectionImplementation implements ClusterConnection, Closeable {
 
   private final RetryingCallerInterceptor interceptor;
 
-  private final OperationInterceptorFactory operationInterceptorFactory;
-
   /**
    * Cluster registry of basic info such as clusterid and meta region location.
    */
@@ -339,7 +337,6 @@ public class ConnectionImplementation implements ClusterConnection, Closeable {
 
     this.stats = ServerStatisticTracker.create(conf);
     this.interceptor = new RetryingCallerInterceptorFactory(conf).build();
-    this.operationInterceptorFactory = createOperationInterceptorFactory(conf);
 
     this.backoffPolicy = ClientBackoffPolicyFactory.create(conf);
 
@@ -373,7 +370,7 @@ public class ConnectionImplementation implements ClusterConnection, Closeable {
         connectionAttributes);
       this.rpcControllerFactory = RpcControllerFactory.instantiate(conf);
       this.rpcCallerFactory = RpcRetryingCallerFactory.instantiate(conf, connectionConfig,
-        interceptor, this.stats, this.metrics, operationInterceptorFactory);
+        interceptor, this.stats, this.metrics);
       this.asyncProcess = new AsyncProcess(this, conf, rpcCallerFactory, rpcControllerFactory);
 
       // Do we publish the status?
@@ -2344,7 +2341,7 @@ public class ConnectionImplementation implements ClusterConnection, Closeable {
   @Override
   public RpcRetryingCallerFactory getNewRpcRetryingCallerFactory(Configuration conf) {
     return RpcRetryingCallerFactory.instantiate(conf, connectionConfig, this.interceptor,
-      this.stats, metrics, createOperationInterceptorFactory(conf));
+      this.stats, metrics);
   }
 
   @Override
@@ -2421,19 +2418,5 @@ public class ConnectionImplementation implements ClusterConnection, Closeable {
       LOG.error("Error fetching cluster ID: ", e);
     }
     return null;
-  }
-
-  private static OperationInterceptorFactory createOperationInterceptorFactory(Configuration conf) {
-    String clazz = conf.get(OperationInterceptorFactory.HBASE_CLIENT_OPERATION_INTERCEPTOR_IMPL);
-    if (clazz == null || clazz.isEmpty()) {
-      return OperationInterceptorFactory.NO_OP;
-    }
-    try {
-      Class<? extends OperationInterceptorFactory> factoryClass =
-        conf.getClassByName(clazz).asSubclass(OperationInterceptorFactory.class);
-      return ReflectionUtils.newInstance(factoryClass, conf);
-    } catch (ClassNotFoundException e) {
-      throw new RuntimeException("Failed to load OperationInterceptorFactory class: " + clazz, e);
-    }
   }
 }
